requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.2
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-js-asset==3.1.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
